requirements.txt
fastapi>=0.110
uvicorn>=0.27
python-multipart>=0.0.9
pillow>=10.0.0
python-docx>=1.1.0
pytesseract>=0.3.10
PyMuPDF>=1.23.7
opencv-python-headless>=4.9.0
